custom jhook now working
need to refine it because it's really
messy to wire up each control at so
many points.

Also need to fox focus ring for
a control that has error as at the
moment it is the dafault gray one
which doesn't look good when you
have jsut moved to a control that
already had an error.