fix: test for submitbutton partially works
the provlem is that render for story only
indicates when it has resumed running the
event handler and not when it has completed.

This means the rest of play function runs
before a re-render has taken place after
submit handler is complete (as the part
of play function ater waiting for event
handler to have resumed goes on task queue
(i.e. microtask queue) whereas between the
end of submit handler and re-render of the
button there is likely a task/event on
job queue involved; perhaps end of
sumit handler is signalled by placing another
event on the job queue). This means we cannot
check if the state of the button has switched
back to being normal.

So I need play to resume only after the
render after sbmit handler has completed.
So I need to convert the signal to a
simple flip-flop in which flop happens
in an effect of the render of the story
rather than immediately after submit
handler has resumed.