fix: further cleanup
problem still the same: if you re-run interaction
test on submit button, it works. but if you go
to the other test of the button and come back to it
it is stuck on loading state. This happens because
after submit handler is exited, submit button
and the other component with useFormStatus that I
have to Flop the flip flop once status is no
longer pending, neither of these is re-rendered.

Somehow it appears that the useFormStatus hook is
not causing a re-render of these two components
within the form.

Perhaps it is losing track of the form's submit
handler?

Also, eliminating args from render and play
functions eliminated the duplicate render
problem