fix: proof flipflop was not the problem
now in main test submit handler only wait for a second
rather than on flip of the flipflop. It worked
fine as long as the second test also used the same
render function and therefore shows its disabled
state for a second before returning to normal.

Now I gave that second test its own render function
(including its own form and rest of JSX) but
in form handler, made it wait for a Promise that
never resolved (a so called floating promise).
This meant that when I went back to first test,
it failed.

Moreover, now going back to that first tests and
re-running the play function did not result in
success unlike the previous commit.