fix: several changes
1. pin daisyui as its higher 5.0..43 version is
causing focus to be lost noce modal alert dialog,
which uses daisyui classes, displayes and sets
focuse to itself.

2. set @swc/helpers as a dev dependency as
for some reason pacakge-lock.json kept
resolving it to 0.5.15 whereas some bundling
(either storybook or next.js build) needed
it to be 0.5.17 or higher.

3. marked .js cnofig files as .mjs
so that next.build doesn't parse them first
as ES6 modules then, when it can't parse them,
reparses/compiles them as CJS modules. This
was inefficient.

4. Upgraded to latest 9.0.12 Storybook