fix: stop auto-rebasing of dependabot PRs
I would like PRs to stay open until it is feasible
to do them and in the mwanwhile new PRs to be generated
for each ecosystem with other updates.

Sometimes it is nto pssible to do an update
e.g. bumping daysui to 5.0.43 led to focus
getting lost on mdal dialogs (which incidentally
used daisyUI classes). I would like to do a bump
no daisyUI boot only when they have release a new
version that fixes the problem. In the meanwhile I
wan the PR to stay open with my cmment describing why
I haven't merged it yet